requirements.txt
pandas>=1.5
numpy
matplotlib
plotly
sqlalchemy
psycopg2-binary
openpyxl
tqdm
